requirements.txt
chromadb==0.4.24
sentence-transformers==2.5.1
numpy==1.26.4
torch==2.2.1
llama-cpp-python>=0.2.0
typer>=0.9.0
rich>=13.0.0
tqdm>=4.65.0
requests>=2.31.0
flask>=2.0.0
flask_sqlalchemy>=3.0.0
flask_cors>=4.0.0 